Extract waitFor test helper to test/wait.ts
Both models/article.background.test.ts and
models/article.lifecycle.test.ts had a private copy of the same
async-polling helper that watches for background-promise side
effects (translation completion, placeholder reset, summary
cleanup) without exposing the underlying promise to the test.
Move it to test/wait.ts so future tests don't grow a third copy
when they hit the same observation pattern.

Behavior is unchanged: same 50 ms poll cadence, same default
10 s deadline, same error message on timeout.

https://github.com/hackers-pub/hackerspub/pull/283#discussion_r3179595557
