+++ models/article.background.test.ts
@@ -11,18 +11,8 @@ import {
   insertAccountWithActor,
   withRollback,
 } from "../test/postgres.ts";
+import { waitFor } from "../test/wait.ts";
 import { generateUuidV7 } from "./uuid.ts";
-
-async function waitFor(predicate: () => Promise<boolean>, timeoutMs = 10000) {
-  const deadline = Date.now() + timeoutMs;
-  while (Date.now() < deadline) {
-    if (await predicate()) return;
-    await new Promise((resolve) => setTimeout(resolve, 50));
-  }
-  throw new Error(
-    `Timed out waiting for async background state after ${timeoutMs}ms`,
-  );
-}
 
 test("startArticleContentSummary() resets summaryStarted when summarization fails", async () => {
   await withRollback(async (tx) => {
